escape_bytes
Function unescape_into

pub fn unescape_into<I>(
    out: &mut [u8],
    i: I,
) -> Result<usize, UnescapeIntoError>
where
    I: IntoIterator,
    I::Item: Borrow<u8>,

Unescape the bytes into the slice.

See crate for the exact rules.

Returns the number of bytes written to the slice.

Errors

If the slice is not large enough to receive the escaped value. No information is provided to support continuing escaping into a new buffer from where it stops. Use the Unescape iterator directly if that is needed.

When encountering unexpected byte sequences.